// src/types.ts
export type Domain = 'mobile' | 'frontend' | 'backend';
export type SkillCategory = 'language' | 'framework' | 'database' | 'tool' | 'other';
export type ProjectContext = 'professional' | 'freelance' | 'personal';
export type ResumeVariant = 'fullstack' | 'mobile' | 'frontend' | 'backend';

export type Me = {
    firstName: string;
    state: string;
    country: string;
    email: string;
    phoneNumber: string;
    github: string;
    linkedin: string;
    twitter: string;
};